ANNOUNCE: linux-0.98 patchlevel 6
You all know what the above subject-line means by now...

Currently only the full sources to 0.98pl6 are available (at
nic.funet.fi in the directory pub/OS/Linux/PEOPLE/Linus), but I'll
probably make context-diffs later today. It depends a bit on the size
of the diffs: the changes in pl6 aren't very suitable for diffing
(renaming of some SCSI files etc).

Anyway, 0.98pl6 is hopefully the last release before 0.99: there are a
few known problems left in this release. Most notable is the serial
code: it works for most people, but others still have problems with it.
I hope this will get fixed within a week (tytso is working on it). It
also seems as if the PS/2 mouse code has some problems.

pl6 contains these fixes:

 - all the tcp/ip patches I've received (and I fixed one bug that
   gcc-2.3 seems to have found).
 - math-emu patch for the problem that resulted in FPU errors with some
   operations.
 - I fixed gcc-2.3 warnings as well as most of the old warnings. You
   shouldn't get more than one or two warnings when recompiling the
   whole kernel.
 - /proc filesystem extensions. Based on ideas (and some code) by
   Darren Senn, but mostly written by yours truly. More about that
   later.
 - some tty_io fixes (there was a bug in the /dev/console handling when
   you changed VC's while using the general console device).
 - re-organization of the keyboard-driver internal data-structures. The
   changes are mostly preliminary: they change the keyboard flags to be
   more easily adaptive to a reprogrammable keyboard driver. No actual
   new features yet.
 - new SCSI drivers: reportedly much faster than the old ones (but not
   all drivers take advantage of it yet..)
 - various other fixes: pty's etc have minor changes.

I hope to make 0.99 in a week or so, and 1.0 after that has been tested
some. I hope people will test out pl6 - 0.99 won't be much different,
and if you don't test pl6, any bugs relating to your particular hardware
may not be found in time for 0.99...

		Linus

===== /proc filesystem comments, ignore if not interested =====

If people want to test out the new /proc filesystem features, please do

  # mount -t proc /proc /proc

(or add the line "/proc /proc proc defaults 0 0" to your /etc/mtab). You
obviously need a /proc subdirectory for the above.

After mounting the proc fs, your /proc should look something like this:

  # ls -l /proc
  ....
  dr-xr-xr-x 4 root root 0 Dec 2 22:55 936
  dr-xr-xr-x 4 root root 0 Dec 2 22:55 983
  -r--r--r-- 1 root root 0 Dec 2 22:55 loadavg
  -r--r--r-- 1 root root 0 Dec 2 22:55 meminfo
  -r--r--r-- 1 root root 0 Dec 2 22:55 uptime

The numeric entries correspond with the pid's of the running procedures,
while the alphabetic entries are vaious general info-files (don't be
fooled by the zero length).

  # cat /proc/meminfo
	  total: used: free: shared: buffers:
  Mem: 15831040 15601664 229376 3952640 8904704
  Swap: 5521408 0 5521408
  # cat /proc/loadavg
  0.01 0.02 0.00
  # cat /proc/uptime
  12981

The info should be pretty self-evident (uptime is in seconds, and yes, I
reboot pretty often).

For each <pid> in the system, the per-process directory looks like this:

  # ls -l /proc/50/
  total 3
  -r--r--r-- 1 root root 0 Dec 2 23:00 cmdline
  lrwx------ 1 root root 3 Dec 2 23:00 cwd -> ---
  -r--r--r-- 1 root root 0 Dec 2 23:00 environ
  lrwx------ 1 root root 3 Dec 2 23:00 exe -> ---
  dr-x------ 2 root root 0 Dec 2 23:00 fd
  dr-x------ 2 root root 0 Dec 2 23:00 lib
  crw------- 1 root root 1, 1 Dec 2 23:00 mem
  lrwx------ 1 root root 3 Dec 2 23:00 root -> ---
  -r--r--r-- 1 root root 0 Dec 2 23:00 stat

The most interesting entries are probably the new ones: cmdline, environ
and stat. They give the process command-line, environment and some
statistics respectively. The file protections will probably have to be
changed.

  # cat /proc/50/cmdline | tr '\000' ' ' && echo
  X :0 -pn
  # cat /proc/50/environ | tr '\000' ' ' && echo
  SHELL=/bin/sh SHLVL=2 BASH=/bin/sh MAIL=/var/spool/mail/root
  HOME=/usr/root PATH=/bin:/usr/bin:/etc:/usr/local/bin:/usr/bin/X11
  LOGNAME=root LESS=-MM TERM=console ignoreeof=10 HOSTTYPE=i386 DISPLAY=:0
  _=/usr/bin/X11/xinit

The 'tr' and 'echo' commands are there to make it look a bit nicer: the
command line arguments and environment variables are separated by '\0'
characters and aren't terminated by a newline. Note that the /proc
filesystem doesn't try to fetch command line data that is swapped out,
so when you are swapping, one or both of the above may be empty or
truncated. Also, the code I wrote to fetch the above may nor be
complete yet, and doesn't seem to be able to find all of the cmdline all
of the time..

Finally there is the (not completely implemented) process status file:

  # cat /proc/50/stat
  50 (X) R 49 50 10 1

This currently gives the pid, short name ("X", max 8 chars, but you get
it even when the process is swapped out), the state (R for running etc),
ppid (49), pgrp (50), session (10) and controlling terminal (1 -
/dev/tty1). The status file will have to be extended to tell you more
about the process, but that should be easy now that the general code is
there..